requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
brotlipy==0.7.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
discord.py==1.7.3
dnspython==2.1.0
docopt==0.6.2
docutils==0.18.1
fett==0.3.2
idna==3.3
multidict==5.2.0
networkx==2.6.3
pkg_resources==0.0.0
pycares==4.1.2
pycparser==2.21
pymongo==3.12.3
python-jsonrpc-server==0.3.4
PyYAML==5.4.1
requests==2.26.0
snooty-lextudio==1.11.5
toml==0.10.2
typing_extensions==4.0.1
ujson==1.35
urllib3==1.26.7
watchdog==1.0.2
yarl==1.7.2
